Migrate pedsum to pedigree-graph 0.8
Pair counts come from estimate_relationship_counts (fold-aware; the exact
set is the six codes it declares, and clamped codes are reported in the
pairs section), per-individual pairs from relationship_pairs, depth from
PedigreeGraph.depth, lineage counts from distinct_ancestor_counts and
descendant_path_counts (n_descendant_paths is now Int64), F from
inbreeding(), and effective size from estimate_effective_sizes. All
eight Ne keys are always present; an unselected or refused estimator
carries a null ne and its reason. Effective-size arrays are indexed by
observed depth label (depths, parent_depths, transition_from/to) and the
transition rates lose their leading null. The EPIMIGHT export reads pair
orientation from block roles and its kinship_exact column carries
float32-origin values. --ne-threads is removed: 0.8 runs the estimators
serially over one prerequisite memo.

Measured on example_pedigree.tsv under --per-individual-pairs the
closest-category fold moves 1C1R 1117->856, 1C2R 774->608, HGAv
517->490, HGGAv 303->256; every other count, the default streaming
counts, and both EPIMIGHT tables are unchanged.

The pin is raised to pedigree-graph>=0.8,<0.9; until the lock is
refreshed, runs use `PYTHONPATH=../pedigree-graph pixi run --frozen`.

+++ tests/test_ancestor_aggregates.py
@@ -43,7 +43,7 @@ def _individual_df() -> pl.DataFrame:
             "n_first_cousins": [0, 0, 0],
             "n_founder_ancestors": np.array([1, 1, 2], dtype=np.int32),
             "n_distinct_ancestors": np.array([0, 0, 2], dtype=np.int32),
-            "n_descendant_paths": np.array([1, 1, 0], dtype=np.int32),
+            "n_descendant_paths": np.array([1, 1, 0], dtype=np.int64),
         }
     )
